pyproject.toml
[tool.poetry]
name = "AAOCASeg"
version = "0.1.1"
description = ""
authors = ["yungselm", "cardionaut", "Pooya Mohammadi Kazaj"]
package-mode = false

[tool.poetry.dependencies]
python = "3.10.12"
loguru = "*"
openpyxl = "*"
matplotlib = "*"
seaborn = "*"
pylint = "*"
hydra-core = "*"
omegaconf = "*"
tensorflow = "2.15.0"
protobuf = ">=3.20.3"
monai = "*"
scikit-image = "*"
pydicom = "*"
customtkinter = "*"
pylibjpeg = "*"
pylibjpeg-libjpeg = "*"
pyqt5 = "5.15.2"
pyqtdarktheme = "2.1.0"
shapely = "*"
SimpleITK = "*"
opencv-python-headless = "^4.9.0.80"
nnunetv2 = {git = "https://github.com/AI-in-Cardiovascular-Medicine/nnUZoo"}
tqdm = "^4.67.1"
torch = "2.4.0"
torchvision = "0.19.0"
pillow = "^11.3.0"
numpy = "^1.26.4"
# opencv-python = "4.8.0.76"

[tool.poetry.group.dev.dependencies]
autohooks = "^23.1.0"
autohooks-plugin-black = "^22.11.0"
autohooks-plugin-isort = "^22.8.0"
pytest = "^7.2.1"
black = "^22.6.0"
isort = "^5.10.1"

[tool.autohooks]
mode = "poetry"
pre-commit = ["autohooks.plugins.black", "autohooks.plugins.isort"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
skip-string-normalization = true

[tool.isort]
profile = "black"


[[tool.poetry.source]]
name = "pytorch_cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"